+++ test/src/test/kotlin/graphql/nadel/tests/next/fixtures/hydration/defer/HydrationDeferInRenamedFieldTestSnapshot.kt
@@ -3,7 +3,7 @@ package graphql.nadel.tests.next.fixtures.hydration.defer
 
 import graphql.nadel.tests.next.ExpectedNadelResponse
 import graphql.nadel.tests.next.ExpectedServiceCall
-import graphql.nadel.tests.next.TestData
+import graphql.nadel.tests.next.TestSnapshot
 import graphql.nadel.tests.next.listOfJsonStrings
 import kotlin.Suppress
 import kotlin.collections.List
@@ -12,39 +12,29 @@ import kotlin.collections.listOf
 /**
  * This class is generated. Do NOT modify.
  *
- * Refer to [graphql.nadel.tests.next.CaptureTestData]
+ * Refer to [graphql.nadel.tests.next.UpdateTestSnapshots
  */
 @Suppress("unused")
-public class HydrationDeferIsDisabledForRelatedIssuesTestData : TestData() {
+public class HydrationDeferInRenamedFieldTestSnapshot : TestSnapshot() {
     override val calls: List<ExpectedServiceCall> = listOf(
             ExpectedServiceCall(
                 service = "issues",
                 query = """
                 | {
-                |   issueByKey(key: "GQLGW-2") {
+                |   rename__issueByKey__getIssueByKey: getIssueByKey(key: "GQLGW-1") {
                 |     key
                 |     hydration__assignee__assigneeId: assigneeId
                 |     __typename__hydration__assignee: __typename
-                |     related {
-                |       hydration__assignee__assigneeId: assigneeId
-                |       __typename__hydration__assignee: __typename
-                |     }
                 |   }
                 | }
                 """.trimMargin(),
                 variables = "{}",
                 response = """
                 | {
-                |   "issueByKey": {
-                |     "key": "GQLGW-2",
-                |     "hydration__assignee__assigneeId": "ari:cloud:identity::user/2",
-                |     "__typename__hydration__assignee": "Issue",
-                |     "related": [
-                |       {
-                |         "hydration__assignee__assigneeId": "ari:cloud:identity::user/1",
-                |         "__typename__hydration__assignee": "Issue"
-                |       }
-                |     ]
+                |   "rename__issueByKey__getIssueByKey": {
+                |     "key": "GQLGW-1",
+                |     "hydration__assignee__assigneeId": "ari:cloud:identity::user/1",
+                |     "__typename__hydration__assignee": "Issue"
                 |   }
                 | }
                 """.trimMargin(),
@@ -71,26 +61,6 @@ public class HydrationDeferIsDisabledForRelatedIssuesTestData : TestData() {
                 delayedResponses = listOfJsonStrings(
                 ),
             ),
-            ExpectedServiceCall(
-                service = "users",
-                query = """
-                | {
-                |   userById(id: "ari:cloud:identity::user/2") {
-                |     name
-                |   }
-                | }
-                """.trimMargin(),
-                variables = "{}",
-                response = """
-                | {
-                |   "userById": {
-                |     "name": "Tom"
-                |   }
-                | }
-                """.trimMargin(),
-                delayedResponses = listOfJsonStrings(
-                ),
-            ),
         )
 
     /**
@@ -98,17 +68,10 @@ public class HydrationDeferIsDisabledForRelatedIssuesTestData : TestData() {
      * {
      *   "data": {
      *     "issueByKey": {
-     *       "key": "GQLGW-2",
-     *       "related": [
-     *         {
-     *           "assignee": {
-     *             "name": "Franklin"
-     *           }
-     *         }
-     *       ],
+     *       "key": "GQLGW-1",
      *       "assignee": {
      *         "value": {
-     *           "name": "Tom"
+     *           "name": "Franklin"
      *         }
      *       }
      *     }
@@ -121,14 +84,7 @@ public class HydrationDeferIsDisabledForRelatedIssuesTestData : TestData() {
             | {
             |   "data": {
             |     "issueByKey": {
-            |       "key": "GQLGW-2",
-            |       "related": [
-            |         {
-            |           "assignee": {
-            |             "name": "Franklin"
-            |           }
-            |         }
-            |       ]
+            |       "key": "GQLGW-1"
             |     }
             |   },
             |   "hasNext": true
@@ -146,7 +102,7 @@ public class HydrationDeferIsDisabledForRelatedIssuesTestData : TestData() {
                 |       ],
                 |       "data": {
                 |         "value": {
-                |           "name": "Tom"
+                |           "name": "Franklin"
                 |         }
                 |       }
                 |     }
