+++ test/src/test/kotlin/graphql/nadel/tests/next/fixtures/hydration/defer/HydrationDeferIsDisabledForRelatedIssuesTestSnapshot.kt
@@ -3,7 +3,7 @@ package graphql.nadel.tests.next.fixtures.hydration.defer
 
 import graphql.nadel.tests.next.ExpectedNadelResponse
 import graphql.nadel.tests.next.ExpectedServiceCall
-import graphql.nadel.tests.next.TestData
+import graphql.nadel.tests.next.TestSnapshot
 import graphql.nadel.tests.next.listOfJsonStrings
 import kotlin.Suppress
 import kotlin.collections.List
@@ -12,61 +12,39 @@ import kotlin.collections.listOf
 /**
  * This class is generated. Do NOT modify.
  *
- * Refer to [graphql.nadel.tests.next.CaptureTestData]
+ * Refer to [graphql.nadel.tests.next.UpdateTestSnapshots
  */
 @Suppress("unused")
-public class HydrationDeferIsDisabledTestData : TestData() {
+public class HydrationDeferIsDisabledForRelatedIssuesTestSnapshot : TestSnapshot() {
     override val calls: List<ExpectedServiceCall> = listOf(
             ExpectedServiceCall(
                 service = "issues",
                 query = """
                 | {
-                |   issues {
+                |   issueByKey(key: "GQLGW-2") {
                 |     key
                 |     hydration__assignee__assigneeId: assigneeId
                 |     __typename__hydration__assignee: __typename
-                |   }
-                | }
-                """.trimMargin(),
-                variables = "{}",
-                response = """
-                | {
-                |   "issues": [
-                |     {
-                |       "key": "GQLGW-1",
-                |       "hydration__assignee__assigneeId": "ari:cloud:identity::user/1",
-                |       "__typename__hydration__assignee": "Issue"
-                |     },
-                |     {
-                |       "key": "GQLGW-2",
-                |       "hydration__assignee__assigneeId": "ari:cloud:identity::user/2",
-                |       "__typename__hydration__assignee": "Issue"
-                |     },
-                |     {
-                |       "key": "GQLGW-3",
-                |       "hydration__assignee__assigneeId": "ari:cloud:identity::user/1",
-                |       "__typename__hydration__assignee": "Issue"
+                |     related {
+                |       hydration__assignee__assigneeId: assigneeId
+                |       __typename__hydration__assignee: __typename
                 |     }
-                |   ]
-                | }
-                """.trimMargin(),
-                delayedResponses = listOfJsonStrings(
-                ),
-            ),
-            ExpectedServiceCall(
-                service = "users",
-                query = """
-                | {
-                |   userById(id: "ari:cloud:identity::user/1") {
-                |     name
                 |   }
                 | }
                 """.trimMargin(),
                 variables = "{}",
                 response = """
                 | {
-                |   "userById": {
-                |     "name": "Franklin"
+                |   "issueByKey": {
+                |     "key": "GQLGW-2",
+                |     "hydration__assignee__assigneeId": "ari:cloud:identity::user/2",
+                |     "__typename__hydration__assignee": "Issue",
+                |     "related": [
+                |       {
+                |         "hydration__assignee__assigneeId": "ari:cloud:identity::user/1",
+                |         "__typename__hydration__assignee": "Issue"
+                |       }
+                |     ]
                 |   }
                 | }
                 """.trimMargin(),
@@ -119,26 +97,21 @@ public class HydrationDeferIsDisabledTestData : TestData() {
      * ```json
      * {
      *   "data": {
-     *     "issues": [
-     *       {
-     *         "key": "GQLGW-1",
-     *         "assignee": {
-     *           "name": "Franklin"
+     *     "issueByKey": {
+     *       "key": "GQLGW-2",
+     *       "related": [
+     *         {
+     *           "assignee": {
+     *             "name": "Franklin"
+     *           }
      *         }
-     *       },
-     *       {
-     *         "key": "GQLGW-2",
-     *         "assignee": {
+     *       ],
+     *       "assignee": {
+     *         "value": {
      *           "name": "Tom"
      *         }
-     *       },
-     *       {
-     *         "key": "GQLGW-3",
-     *         "assignee": {
-     *           "name": "Franklin"
-     *         }
      *       }
-     *     ]
+     *     }
      *   }
      * }
      * ```
@@ -147,30 +120,39 @@ public class HydrationDeferIsDisabledTestData : TestData() {
             response = """
             | {
             |   "data": {
-            |     "issues": [
-            |       {
-            |         "key": "GQLGW-1",
-            |         "assignee": {
-            |           "name": "Franklin"
+            |     "issueByKey": {
+            |       "key": "GQLGW-2",
+            |       "related": [
+            |         {
+            |           "assignee": {
+            |             "name": "Franklin"
+            |           }
             |         }
-            |       },
-            |       {
-            |         "key": "GQLGW-2",
-            |         "assignee": {
-            |           "name": "Tom"
-            |         }
-            |       },
-            |       {
-            |         "key": "GQLGW-3",
-            |         "assignee": {
-            |           "name": "Franklin"
-            |         }
-            |       }
-            |     ]
-            |   }
+            |       ]
+            |     }
+            |   },
+            |   "hasNext": true
             | }
             """.trimMargin(),
             delayedResponses = listOfJsonStrings(
+                """
+                | {
+                |   "hasNext": false,
+                |   "incremental": [
+                |     {
+                |       "path": [
+                |         "issueByKey",
+                |         "assignee"
+                |       ],
+                |       "data": {
+                |         "value": {
+                |           "name": "Tom"
+                |         }
+                |       }
+                |     }
+                |   ]
+                | }
+                """.trimMargin(),
             ),
         )
 }
